clean up util
